Implement claiming logic

+++ src/routes/index.svelte
@@ -4,6 +4,7 @@
 	import users from '$lib/data/users.js';
 	import tempItems from '$lib/data/items.js';
 
+	const userId = 1
 	let items = tempItems;
 
 	const addItem = (event) => {
@@ -13,12 +14,16 @@
 		items = [...items];
 	};
 	const claimItem = (event) => {
-		console.log(`Received claim event ${event.detail}`);
+		console.log(`Received claim event for item ${event.detail}`);
 		items = items.map((item) => {
 			if (item.id === event.detail) {
-				item.claimed = !item.claimed;
-				item.claimedBy = [...item.claimedBy, 1];
-				console.log(item);
+				const claims = [...item.claimedBy]
+				if (claims.includes(userId)) {
+					let newClaims = claims.filter(i => i !== userId)
+					item.claimedBy = new Set(newClaims)
+				} else {
+					item.claimedBy = new Set([...claims, userId]);
+				}
 			}
 			return item;
 		});
@@ -27,12 +32,13 @@
 		console.log('Received delete event');
 		items = items.filter((item) => item.id !== event.detail);
 	};
+
 </script>
 
 <main class="bg-base-100 py-4 px-2 md:px-6 min-h-screen">
 	<h1 class="text-blue-600 font-thin text-3xl my-4 text-center">Ich packe meinen Koffer...</h1>
 	<div class="divider font-extralight text-3xl text-neutral-content">~</div>
-	<List {items} on:claim={claimItem} on:delete={deleteItem} />
+	<List {items} {userId} on:claim={claimItem} on:delete={deleteItem} />
 
 	<div class="divider font-extralight text-3xl text-neutral-content">~</div>
 	<AddItemForm on:addItem={addItem} />
